tests: Simplify EK tests
Merge the three tests into a template test that CMake can configure.
Remove the `eval()`-based re-implementation of the Python dict.
