+++ testsuite/python/test_checkpoint.py
@@ -107,25 +107,14 @@ class CheckpointTest(ut.TestCase):
             self.assertIn(key, state)
             self.assertAlmostEqual(reference[key], state[key], delta=1E-5)
         state_species = ek_species.get_params()
-        reference_species = {'density': 0.4, 'D': 0.02, 'valency': 0.3}
+        reference_species = {'density': 0.4, 'D': 0.02, 'valency': 0.3,
+                             'ext_force_density': [0.01, -0.08, 0.06]}
         for key in reference_species:
             self.assertIn(key, state_species)
-            self.assertAlmostEqual(
+            np.testing.assert_allclose(
                 reference_species[key],
                 state_species[key],
-                delta=1E-5)
-        self.assertAlmostEqual(
-            state_species['ext_force_density'][0],
-            0.01,
-            delta=1E-5)
-        self.assertAlmostEqual(
-            state_species['ext_force_density'][1],
-            -0.08,
-            delta=1E-5)
-        self.assertAlmostEqual(
-            state_species['ext_force_density'][2],
-            0.06,
-            delta=1E-5)
+                atol=1E-5)
 
     def test_variables(self):
         self.assertEqual(system.cell_system.skin, 0.1)
